random done

+++ app.rb
@@ -41,8 +41,8 @@ end
 
 get("/random/results") do
 
-  @the_min_num = params.fetch("users_min_number").to_i
-  @the_max_num = params.fetch("users_max_number").to_i
+  @the_min_num = params.fetch("users_min_number").to_f
+  @the_max_num = params.fetch("users_max_number").to_f
   @the_result = rand(@the_min_num..@the_max_num)
 
   erb(:random_results)
